+++ app/Cargo.toml
@@ -73,7 +73,7 @@ bincode.workspace = true
 bitflags.workspace = true
 bitflags-serde-legacy.workspace = true
 blocking.workspace = true
-bounded-vec-deque = "0.1"
+bounded-vec-deque.workspace = true
 bytecount = "0.6.9"
 bytes.workspace = true
 byte-unit = "5.1.4"
