+++ crates/remote_server/Cargo.toml
@@ -9,6 +9,7 @@ license.workspace = true
 [dependencies]
 anyhow.workspace = true
 async-channel.workspace = true
+bounded-vec-deque.workspace = true
 command.workspace = true
 dashmap.workspace = true
 futures.workspace = true
