+++ crates/warpui_core/Cargo.toml
@@ -25,7 +25,7 @@ async-channel.workspace = true
 async-executor = "1.4.1"
 async-task.workspace = true
 async-fs.workspace = true
-bounded-vec-deque = "0.1"
+bounded-vec-deque.workspace = true
 bytes.workspace = true
 cfg-if.workspace = true
 chrono = "0.4"
